requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
datasets>=2.14.0
huggingface_hub>=0.17.0
scikit-learn>=1.3.0
xgboost>=1.7.0
optuna>=3.4.0
wandb>=0.15.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pytz>=2023.3
pvlib>=0.10.0
scipy>=1.11.0
matplotlib>=3.7.0
tqdm>=4.66.0
pyyaml>=6.0
icecream>=2.1.0
joblib>=1.3.0